refactor: update __init__.py to import from new subpackages and remove figure_plots.py shim
Update top-level __init__.py to import directly from axes, plot1d, plot2d,
and plot3d subpackages instead of the figure_plots compat shim. Delete
figure_plots.py and fix all test files that still imported from it.
